+++ packages/core/src/components/NumberField/__stories__/NumberField.stories.tsx
@@ -12,7 +12,7 @@ const metaSettings = createStoryMetaSettingsDecorator({
 });
 
 export default {
-  title: "Components/NumberField",
+  title: "Components/NumberField [New]",
   component: NumberField,
   argTypes: metaSettings.argTypes,
   decorators: metaSettings.decorators