+++ packages/core/src/components/index.ts
@@ -30,6 +30,7 @@ export * from "./Heading";
 export * from "./HiddenText";
 export * from "@vibe/icon";
 export * from "./IconButton";
+export * from "./Info";
 export * from "./Label";
 export * from "./LayerProvider";
 export * from "./Link";